+++ db/migrate/20151211193840_create_supplier_point_products.rb
@@ -3,7 +3,7 @@ class CreateSupplierPointProducts < ActiveRecord::Migration
     create_table :supplier_point_products do |t|
       t.belongs_to :supplier, index: true, foreign_key: true
       t.belongs_to :product, index: true, foreign_key: true
-      t.integer :points, default: 0
+      t.float :points, default: 0
 
       t.timestamps null: false
     end
